testing/web-platform/mozilla/tests/css/css-grid/subgrid/subgrid-baseline-008.html

<!DOCTYPE html>
<!-- TODO: Remove this test once bug 1871719 is addressed. This test is just a
     modified copy of a more-robust WPT test, with several forms of padding
     removed to provide test coverage while working around bug 1871719. -->
<link rel="help" href="https://drafts.csswg.org/css-grid-2/">
<style>
.item {
  inline-size: 40px;
  box-sizing: border-box;
  border: solid 5px hotpink;
  line-height: 0;
  margin-block-start: 3px;
  margin-block-end: 5px;
}
.small {
  width: 20px;
  height: 20px;
  border: solid 5px cyan;
}
.first {
  align-self: baseline;
}
.last {
  align-self: last baseline;
}
.item.small.first {
  block-size: 50px;
}
.item.small.last {
  block-size: 100px;
}
span {
  width: 20px;
  height: 20px;
  box-sizing: border-box;
  border: solid 5px orange;
  display: inline-block;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.item')">

<div style="writing-mode: vertical-rl; width: 600px; position: relative; display: grid; grid-template: auto auto auto / 100px 100px 100px 100px;">
  <div style="display: grid;
              grid-column: 1 / span 2;
              grid-row: 1 / span 3;
              grid-template: subgrid / subgrid;">
    <div style="display: grid;
                grid-column: 1 / span 2;
                grid-row: 1 / span 2;
                grid-template: subgrid / subgrid;">
      <div data-offset-x="537" class="item first">
        <span></span><br><span></span>
      </div>
      <div data-offset-x="440" class="item last">
        <span></span><br><span></span>
      </div>
      <div data-offset-x="337" class="item first">
        <span></span><br><span></span>
      </div>
      <div data-offset-x="240" class="item last">
        <span></span><br><span></span>
      </div>
    </div>
    <div data-offset-x="137" class="item first">
      <span></span><br><span></span>
    </div>
    <div data-offset-x="40" class="item last">
      <span></span><br><span></span>
    </div>
  </div>
  <div data-offset-x="547" class="item small first"></div>
  <div data-offset-x="405" class="item small last"></div>
  <div data-offset-x="347" class="item small first"></div>
  <div data-offset-x="205" class="item small last"></div>
  <div data-offset-x="147" class="item small first"></div>
  <div data-offset-x="5" class="item small last"></div>
</div>
